refactoring html to modular components
