+++ src/app/components/RestaurantCard.tsx
@@ -1,0 +1,29 @@
+import Link from 'next/link'
+import React from 'react'
+
+const RestaurantCard = () => {
+  return (
+    <div className="w-64 h-72 m-3 rounded overflow-hidden border cursor-pointer">
+      <Link href="/restaurant/milestones-grill">
+        <img src="https://resizer.otstatic.com/v2/photos/wide-huge/2/31852905.jpg" alt="" className="w-full h-36" />
+        <div className="p-1">
+          <h3 className="font-bold text-2xl mb-2">Milestones Grill</h3>
+          <div className="flex items-start">
+            <div className="flex mb-2">
+              ****
+            </div>
+            <p className="ml-2">77 Reviews</p>
+          </div>
+          <div className="flex text-reg font-light">
+            <p className="mr-3 capitalize">Mexican</p>
+            <p className="mr-3">$$$$</p>
+            <p>Toronto</p>
+          </div>
+          <p className="text-sm mt-1 font-bold">Booked 3 times today</p>
+        </div>
+      </Link>
+    </div>
+  )
+}
+
+export default RestaurantCard
