+++ src/app/restaurant/[slug]/components/Header.tsx
@@ -1,0 +1,17 @@
+import React from 'react'
+
+const Header = () => {
+  return (
+  <div className="h-96 overflow-hidden">
+    <div
+      className="bg-center bg-gradient-to-r from-[#0f1f47] to-[#5f6984] h-full flex justify-center items-center"
+    >
+      <h1 className="text-7xl text-white captitalize text-shadow text-center">
+        Milestones Grill (Toronto)
+      </h1>
+    </div>
+  </div>
+  )
+}
+
+export default Header;
