+++ src/app/restaurant/[slug]/components/Menu.tsx
@@ -1,0 +1,20 @@
+import MenuCard from "./MenuCard";
+
+const Menu = () => {
+  return (
+      <main className="bg-white mt-5">
+      <div>
+        <div className="mt-4 pb-1 mb-1">
+          <h1 className="font-bold text-4xl">
+            Menu
+          </h1>
+          <div className="flex flex-wrap justify-between">
+            <MenuCard />
+          </div>
+        </div>
+      </div>
+    </main>
+  )
+}
+
+export default Menu;
