+++ src/app/restaurant/[slug]/components/RestaurantNavBar.tsx
@@ -1,0 +1,13 @@
+import Link from 'next/link'
+import React from 'react'
+
+const RestaurantNavBar = () => {
+  return (
+    <nav className="flex text-reg border-b border-gray-500 pb-2">
+      <Link href="/restaurant/milestones-grill" className="mr-7">Overview</Link>
+      <Link href="/restaurant/milestones-grill/menu" className="mr-7">Menu</Link>
+    </nav>
+  )
+}
+
+export default RestaurantNavBar;
