+++ src/app/restaurant/[slug]/components/Title.tsx
@@ -1,0 +1,9 @@
+const Title = () => {
+  return (
+    <div className="mt-4 border-b border-gray-500 pb-6">
+      <h1 className="font-bold text-6xl">Milestones Grill</h1>
+    </div>
+  )
+}
+
+export default Title;
